velion-backend/django/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
drf-yasg==1.21.5
dj-database-url==2.0.0
psycopg2-binary==2.9.6
neo4j==5.8.0
redis==4.5.5
python-dotenv==1.0.0
gunicorn==20.1.0
whitenoise==6.4.0
pillow==9.5.0
celery==5.2.7
djangorestframework-simplejwt==5.2.2
django-extensions==3.2.3
